:truck: Move mandala chart section to reuse
