+++ packages/web/gql/gql.ts
@@ -13,6 +13,7 @@ import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/
  * Therefore it is highly recommended to use the babel or swc plugin for production.
  */
 const documents = {
+    "\n  query GetRecommendationForSubGoals(\n    $mainGoal: String!\n    $selectedSubGoals: [String!]\n    $currentLanguage: String!\n  ) {\n    recommendationForSubGoals(\n      mainGoal: $mainGoal\n      selectedSubGoals: $selectedSubGoals\n      currentLanguage: $currentLanguage\n    ) {\n      ... on RecommendationSuccess {\n        recommendations {\n          text\n        }\n      }\n      ... on RecommendationFailure {\n        errorType\n      }\n    }\n  }\n": types.GetRecommendationForSubGoalsDocument,
     "\n  mutation SignOut {\n    signOut\n  }\n": types.SignOutDocument,
     "\n  mutation DeleteAccount($email: String!) {\n    deleteAccount(email: $email) {\n      ... on DeleteAccountSuccess {\n        success\n      }\n      ... on DeleteAccountFailure {\n        errorType\n      }\n    }\n  }\n": types.DeleteAccountDocument,
     "\n  mutation FindPassword($email: String!) {\n    findPassword(email: $email) {\n      ... on FindPasswordSuccess {\n        success\n      }\n      ... on FindPasswordFailure {\n        errorType\n      }\n    }\n  }\n": types.FindPasswordDocument,
@@ -20,7 +21,6 @@ const documents = {
     "\n  mutation SignIn(\n    $email: String!\n    $password: String!\n    $keepSignedIn: Boolean!\n  ) {\n    signIn(email: $email, password: $password, keepSignedIn: $keepSignedIn) {\n      ... on SignInSuccess {\n        token\n        user {\n          _id\n          email\n          nickname\n          createdAt\n          emailVerification {\n            isVerified\n            token\n            expiresAt\n          }\n          resetPassword {\n            token\n            expiresAt\n            isVerified\n          }\n          purchasedInfo {\n            isPurchased\n            purchasedAt\n            expiresAt\n          }\n        }\n      }\n      ... on SignInFailure {\n        errorType\n      }\n    }\n  }\n": types.SignInDocument,
     "\n  mutation SignUp(\n    $email: String!\n    $password: String!\n    $passwordConfirm: String!\n    $nickname: String!\n  ) {\n    signUp(\n      email: $email\n      password: $password\n      passwordConfirm: $passwordConfirm\n      nickname: $nickname\n    ) {\n      ... on SignUpSuccess {\n        isMailSent\n      }\n      ... on SignUpFailure {\n        errorType\n      }\n    }\n  }\n": types.SignUpDocument,
     "\n  mutation VerifyEmail($token: String!) {\n    verifyEmail(token: $token) {\n      ... on VerifyEmailSuccess {\n        success\n      }\n      ... on VerifyEmailFailure {\n        errorType\n      }\n    }\n  }\n": types.VerifyEmailDocument,
-    "\n  query GetRecommendationForSubGoals(\n    $mainGoal: String!\n    $selectedSubGoals: [String!]\n    $currentLanguage: String!\n  ) {\n    recommendationForSubGoals(\n      mainGoal: $mainGoal\n      selectedSubGoals: $selectedSubGoals\n      currentLanguage: $currentLanguage\n    ) {\n      ... on RecommendationSuccess {\n        recommendations {\n          text\n        }\n      }\n      ... on RecommendationFailure {\n        errorType\n      }\n    }\n  }\n": types.GetRecommendationForSubGoalsDocument,
 };
 
 /**
@@ -37,6 +37,10 @@ const documents = {
  */
 export function graphql(source: string): unknown;
 
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query GetRecommendationForSubGoals(\n    $mainGoal: String!\n    $selectedSubGoals: [String!]\n    $currentLanguage: String!\n  ) {\n    recommendationForSubGoals(\n      mainGoal: $mainGoal\n      selectedSubGoals: $selectedSubGoals\n      currentLanguage: $currentLanguage\n    ) {\n      ... on RecommendationSuccess {\n        recommendations {\n          text\n        }\n      }\n      ... on RecommendationFailure {\n        errorType\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetRecommendationForSubGoals(\n    $mainGoal: String!\n    $selectedSubGoals: [String!]\n    $currentLanguage: String!\n  ) {\n    recommendationForSubGoals(\n      mainGoal: $mainGoal\n      selectedSubGoals: $selectedSubGoals\n      currentLanguage: $currentLanguage\n    ) {\n      ... on RecommendationSuccess {\n        recommendations {\n          text\n        }\n      }\n      ... on RecommendationFailure {\n        errorType\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
@@ -65,10 +69,6 @@ export function graphql(source: "\n  mutation SignUp(\n    $email: String!\n    
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  mutation VerifyEmail($token: String!) {\n    verifyEmail(token: $token) {\n      ... on VerifyEmailSuccess {\n        success\n      }\n      ... on VerifyEmailFailure {\n        errorType\n      }\n    }\n  }\n"): (typeof documents)["\n  mutation VerifyEmail($token: String!) {\n    verifyEmail(token: $token) {\n      ... on VerifyEmailSuccess {\n        success\n      }\n      ... on VerifyEmailFailure {\n        errorType\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query GetRecommendationForSubGoals(\n    $mainGoal: String!\n    $selectedSubGoals: [String!]\n    $currentLanguage: String!\n  ) {\n    recommendationForSubGoals(\n      mainGoal: $mainGoal\n      selectedSubGoals: $selectedSubGoals\n      currentLanguage: $currentLanguage\n    ) {\n      ... on RecommendationSuccess {\n        recommendations {\n          text\n        }\n      }\n      ... on RecommendationFailure {\n        errorType\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetRecommendationForSubGoals(\n    $mainGoal: String!\n    $selectedSubGoals: [String!]\n    $currentLanguage: String!\n  ) {\n    recommendationForSubGoals(\n      mainGoal: $mainGoal\n      selectedSubGoals: $selectedSubGoals\n      currentLanguage: $currentLanguage\n    ) {\n      ... on RecommendationSuccess {\n        recommendations {\n          text\n        }\n      }\n      ... on RecommendationFailure {\n        errorType\n      }\n    }\n  }\n"];
 
 export function graphql(source: string) {
   return (documents as any)[source] ?? {};
