+++ packages/web/src/hooks/useMandalaChart.tsx/useMandalaChart.i18n.ts
@@ -1,0 +1,19 @@
+const TRANSLATIONS = {
+  en: {
+    aiModeDescription: `You cannot type in AI mode. Touch the recommended words to add them to the Mandala Chart.`,
+    downloadImage: 'Download Chart',
+    saveChart: 'Save Chart',
+  },
+  ko: {
+    aiModeDescription: `AI 모드에서는 직접 입력할 수 없습니다. 추천 단어를 선택해 만다라트 계획표에 추가해주세요.`,
+    downloadImage: '이미지 다운로드',
+    saveChart: '차트 저장',
+  },
+  'zh-Hant': {
+    aiModeDescription: `AI 模式下無法直接輸入。請選擇推薦的文字，加入曼陀羅思考法九宮格。`,
+    downloadImage: '下載圖表',
+    saveChart: '保存圖表',
+  },
+}
+
+export default TRANSLATIONS
