+++ packages/web/src/pages/index.i18n.ts
@@ -5,9 +5,6 @@ const TRANSLATIONS = {
     description: `The Mandala Chart is a powerful tool
     for visualizing and achieving goals. It helps in organizing thoughts and
     actions towards a specific objective.`,
-    aiModeDescription: `You cannot type in AI mode. Touch the recommended words to add them to the Mandala Chart.`,
-    downloadImage: 'Download Chart',
-    saveChart: 'Save Chart',
     loadSavedChart: "If you've saved your chart, click here to load.",
   },
   ko: {
@@ -15,18 +12,12 @@ const TRANSLATIONS = {
     title: '만다라트 계획표',
     description: `만다라트 계획표는 목표를 성취하기 위한 강력한 도구입니다. 특정 목표를 위한
     생각과 행동을 조직화하는 데 도움이 됩니다.`,
-    aiModeDescription: `AI 모드에서는 직접 입력할 수 없습니다. 추천 단어를 선택해 만다라트 계획표에 추가해주세요.`,
-    downloadImage: '이미지 다운로드',
-    saveChart: '차트 저장',
     loadSavedChart: '저장한 차트가 있다면 여기를 클릭하여 불러오세요.',
   },
   'zh-Hant': {
     tabTitle: 'Life Leader | 制定自己的曼陀羅思考法九宮格',
     title: '曼陀羅思考法九宮格',
     description: `曼陀羅思考法九宮格令夢想變得圖像化，助你早日達成理想。這種思考法可以用作組織思考和行動，一步步具體實現目標。`,
-    aiModeDescription: `AI 模式下無法直接輸入。請選擇推薦的文字，加入曼陀羅思考法九宮格。`,
-    downloadImage: '下載圖表',
-    saveChart: '保存圖表',
     loadSavedChart: '如果您已經保存了圖表，請點擊此處加載。',
   },
 }
